move mediamonitor from libmyth to libmythui (needs distclean)
