+++ mythtv/libs/libmyth/CMakeLists.txt
@@ -15,7 +15,6 @@ set(LIBMYTH_HEADERS
     mythaverror.h
     mythcontext.h
     mythexp.h
-    mythmediamonitor.h
     output.h
     visual.h
     )
@@ -70,7 +69,6 @@ add_library(
   dbsettings.cpp
   mythaverror.cpp
   mythcontext.cpp
-  mythmediamonitor.cpp
   output.cpp
   )
 
@@ -96,18 +94,11 @@ if(TARGET PkgConfig::PULSEAUDIO)
                                 audio/audiooutputpulse.h)
   endif()
 endif()
-if(UNIX)
-  target_sources(myth PRIVATE mediamonitor-unix.cpp mediamonitor-unix.h)
-  if(NOT TARGET ${QT_PKG_NAME}::DBus)
-    set_source_files_properties(mediamonitor-unix.h PROPERTIES SKIP_AUTOMOC
-                                                               TRUE)
-  endif()
-endif()
+
 if(WIN32)
   target_sources(
     myth
-    PRIVATE mediamonitor-windows.cpp mediamonitor-windows.h
-            audio/audiooutputwin.cpp audio/audiooutputdx.cpp
+    PRIVATE audio/audiooutputwin.cpp audio/audiooutputdx.cpp
             audio/audiooutputwin.h audio/audiooutputdx.h)
 elseif(ANDROID)
   target_sources(
@@ -116,7 +107,7 @@ elseif(ANDROID)
 endif()
 
 if(MINGW OR win32-msvc)
-  target_sources(myth PRIVATE mediamonitor-windows.cpp audio/audiooutputwin.cpp
+  target_sources(myth PRIVATE audio/audiooutputwin.cpp
                               audio/audiooutputdx.cpp)
   target_link_libraries(myth PRIVATE winmm ws2_32 user32 PkgConfig::SAMPLERATE
                                      PkgConfig::SoundTouch)
@@ -127,13 +118,7 @@ if(APPLE)
   target_link_libraries(
     myth
     PRIVATE ${APPLE_APPLICATIONSERVICES_LIBRARY} ${APPLE_AUDIOTOOLBOX_LIBRARY}
-            ${APPLE_AUDIOUNIT_LIBRARY} ${APPLE_COREAUDIO_LIBRARY}
-            ${APPLE_IOKIT_LIBRARY})
-  if(DARWIN_DA)
-    set_target_properties(myth PROPERTIES COMPILE_DEFINITIONS USING_DARWIN_DA)
-    target_sources(myth PRIVATE mediamonitor-darwin.cpp mediamonitor-darwin.h)
-    target_link_libraries(myth PRIVATE ${APPLE_DISKARBITRATION_LIBRARY})
-  endif()
+            ${APPLE_AUDIOUNIT_LIBRARY} ${APPLE_COREAUDIO_LIBRARY})
 endif()
 
 #
@@ -160,7 +145,6 @@ target_link_libraries(
           $<TARGET_NAME_IF_EXISTS:PkgConfig::ALSA>
           $<TARGET_NAME_IF_EXISTS:PkgConfig::JACK>
           $<TARGET_NAME_IF_EXISTS:PkgConfig::PULSEAUDIO>
-          $<TARGET_NAME_IF_EXISTS:Qt${QT_VERSION_MAJOR}::DBus>
           $<$<BOOL:${WIN32}>:winmm>
           $<$<BOOL:${WIN32}>:ws2_32>
           $<$<BOOL:${WIN32}>:user32>
