+++ mythtv/libs/libmythui/mediamonitor-darwin.cpp
@@ -13,7 +13,7 @@
 #include "libmythbase/mythlogging.h"
 
 #include "mediamonitor-darwin.h"
-#include "mythmediamonitor.h"
+#include "mediamonitor.h"
 
 #include <IOKit/IOKitLib.h>
 #include <IOKit/storage/IOMedia.h>
