+++ mythtv/libs/libmythui/mediamonitor-windows.cpp
